Isolate the growth test
